+++ xbird-cloud-netflix/xbird-cloud-netflix-ribbon/src/main/java/com/xbird/cloud/netflix/ribbon/RibbonApplication.java
@@ -1,0 +1,40 @@
+package com.xbird.cloud.netflix.ribbon;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.loadbalancer.LoadBalanced;
+import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
+import org.springframework.cloud.netflix.hystrix.EnableHystrix;
+import org.springframework.cloud.netflix.hystrix.dashboard.EnableHystrixDashboard;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.client.RestTemplate;
+
+@SpringBootApplication
+@EnableEurekaClient
+@EnableHystrix
+@EnableHystrixDashboard
+@RestController
+public class RibbonApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(RibbonApplication.class, args);
+  }
+  
+  @Bean
+  @LoadBalanced
+  public RestTemplate restTemplate() {
+    return new RestTemplate();
+  }
+  
+  @Autowired
+  private HelloService helloService;
+
+  @RequestMapping(value = "/")
+  public String hi(@RequestParam(defaultValue = "World") String name) {
+    return helloService.hiService(name);
+  }
+}