+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - [新增] 添加 Spring Cloud OpenFeign 项目。
 - [新增] 添加 Spring Cloud Netflix Hystrix 项目。
 - [新增] 添加 Spring Cloud Netfilx Turbine 项目。
+- [新增] 添加 Spring Cloud Netfilx Ribbon 项目。